+++ cmd/alpine/main.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	"github.com/csmith/envflag/v2"
-	"github.com/csmith/latest/v2"
+	"github.com/csmith/latest/v3"
 )
 
 var showJson = flag.Bool("json", false, "Provide output in json")