+++ cmd/alpinepackage/main.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"github.com/csmith/envflag/v2"
-	"github.com/csmith/latest/v2"
+	"github.com/csmith/latest/v3"
 )
 
 var showJson = flag.Bool("json", false, "Provide output in json")