+++ cmd/alpinepackagelist/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/csmith/apkutils/v2"
 	"github.com/csmith/apkutils/v2/keys"
 	"github.com/csmith/envflag/v2"
-	"github.com/csmith/latest/v2"
+	"github.com/csmith/latest/v3"
 )
 
 var showJson = flag.Bool("json", false, "Provide output in json")